Update test case image and docstring

+++ main.py
@@ -5,9 +5,13 @@ from prettytable import PrettyTable
 from prettytable import from_db_cursor
 
 list_item = []
-engine = create_engine('sqlite:///transaction.db')
+engine = create_engine('sqlite:///transactions.db')
+
 
 def create_table():
+    """
+    Function untuk men-create table transaction
+    """
     
     conn = engine.connect()
 
@@ -24,15 +28,16 @@ harga_diskon float
     conn.execute(query)
     query = text("""SELECT * FROM transactions""")
     results = conn.execute(query)
-    for result in results:
-        print(result)
 
     conn.close()
 
-def insert_to_table():
-    
-    for item in list_item:
 
+#
+def insert_to_table():
+    """
+    Function untuk memasukkan list transaksi ke dalam table transaction.
+    """
+    for item in list_item:
         list_to_dic = {
             'nama_item' : item[0], 
             'jumlah_item' : item[1], 
@@ -49,13 +54,12 @@ def insert_to_table():
             conn.execute(sql, list_to_dic)
             conn.commit()
         conn.close()
-        
-    
-        
-
 
 
 def print_pesanan():
+    """
+    Function untuk menampilkan pesanan yang ada di list.
+    """
     table_view = PrettyTable()
 
     table_view.field_names = ["Nama Item", "Jumlah Item", "Harga/Item", "Total Harga"]
@@ -67,19 +71,29 @@ def print_pesanan():
 
 
 def check_diskon(total_harga):
+    """
+    Function untuk melakukan check diskon yang didapatkan oleh item.
+
+    Parameter:
+        total_harga (float): total harga per item (item quantity * item price).
+
+    Returns:
+        diskon (float): harga diskon yang didapatkan per item
+        harga_diskon (float): total harga per item dikurangi diskon
+    """
     if((total_harga) >= float(200000)):
         diskon = total_harga * 0.05
         harga_diskon = total_harga * 0.95
         
         return diskon, harga_diskon
     
-    elif((total_harga) >= 300000):
+    elif((total_harga) >= float(300000):
         diskon = total_harga * 0.06
         harga_diskon = total_harga * 0.94
         
         return diskon, harga_diskon
     
-    elif((total_harga) >= 500000):
+    elif((total_harga) >= float(500000):
         diskon = total_harga * 0.07
         harga_diskon = total_harga * 0.93
         
@@ -90,16 +104,26 @@ def check_diskon(total_harga):
         return diskon, harga_diskon
 
 
-##function to add master item 
 def add_item(item_name, item_qty, price_per_item):
+    """
+    Function untuk menambahkan item kedalam list.
+
+    Parameter:
+        item_name (string): nama dari item.
+        item_qty (float): quantity dari item
+        price_per_item (float): harga unit dari item.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
     total_harga = item_qty * price_per_item
 
-    #check if name is empty
+    #check nama item sudah diisi
     if(len(item_name) == 0):
         print("Nama barang tidak boleh kosong")
         menu()
 
-    #check apakah item sudah terdaftar
+    #check item sudah terdaftar
     isExist = any(item_name.lower() 
                     in item[0].lower() 
                     for item in list_item) 
@@ -117,10 +141,21 @@ def add_item(item_name, item_qty, price_per_item):
               
     return list_item
 
-##function to update name of an item  
+
 def update_item_name(item_name, new_item_name):
+    """
+    Function untuk mengupdate nama dari item yang berada di list.
+
+    Parameter:
+        item_name (string): nama dari item.
+        new_item_name (string): nama item yang baru.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
+
     for item_list in list_item:
-        if item_list[0] == item_name:
+        if item_list[0].lower()  == item_name.lower():
             item_list[0] = new_item_name
             print(f'''
 item name: {item_list[0]}
@@ -132,9 +167,20 @@ item price: {item_list[2]}
 
     return list_item
     
-
-##function to update qty of an item   
+ 
 def update_item_qty(item_name, new_item_qty):
+    """
+    Function untuk mengupdate quantity dari item yang berada di list.
+
+    Parameter:
+        item_name (string): nama dari item.
+        new_item_qty (string): quantity item yang baru.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
+    
+
     for item_list in list_item:
         if item_list[0] == item_name:
             item_list[1] = new_item_qty
@@ -147,8 +193,18 @@ item price: {item_list[2]}''')
 
     return list_item
 
-##function to update price of an item   
+  
 def update_item_price(item_name, new_item_price):
+    """
+    Function untuk mengupdate unit price dari item yang berada di list.
+
+    Parameter:
+        item_name (string): nama dari item.
+        new_item_price (string): unit price item yang baru.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
     for item_list in list_item:
         if item_list[0] == item_name:
             item_list[2] = new_item_price
@@ -161,35 +217,60 @@ item price: {item_list[2]}''')
 
     return list_item
 
-##function to delete  an item   
+   
 def delete_item(item_name):
+    """
+    Function untuk menghapus item yang berada di list.
+
+    Parameter:
+        item_name (string): nama dari item.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
+    popped_item = None
     for i in range(len(list_item)):
         if list_item[i][0] == item_name:
             popped_item = list_item.pop(i)
-            print(popped_item)
+            print_pesanan()
 
+    if (popped_item is None):
+        print('Item tidak ditemukan.') 
+               
     menu()
     return list_item
           
 
 
-##function to reset an transaction 
 def reset_transaction():
+    """
+    Function untuk menghapus semua item yang berada di list.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
     reset = input("Are you sure you want to delete all item(Y/N): ")
-    if reset == 'Y' :
+    if reset.upper() == 'Y' :
         list_item.clear()
+        print('Semua item berhasil dihapus.')
         menu()
-    elif reset == 'N':
+    elif reset.upper() == 'N':
         menu()
     else :
         print("Invalid input. Enter Y (Yes) or N (No)")
         reset_transaction()
+        
 
     return list_item
 
-    
-##function to check an transaction
+
 def check_order():
+    """
+    Function untuk menghapus semua item yang berada di list.
+
+    Returns:
+        list_item (list) = item yang sudah dimasukkan.
+    """
     total_bayar = 0
 
     ##chack if item quantity and price is null or negatif
@@ -207,11 +288,15 @@ def check_order():
 
     return list_item 
 
-##function to check-out an transaction
+
 def check_out():
+    """
+    Function untuk melakukan check-out terhadap item-item yang ada di list.
+    """
     create_table()
     insert_to_table()
-    print_pesanan()
+
+    total_bayar = 0.0
 
     for item in list_item:
         total_bayar = total_bayar + item[5]
@@ -227,6 +312,13 @@ def check_out():
     
     print(f"Total yang harus Dibayar: {total_bayar} ")
 
+def leave():
+    """
+    Function untuk keluar dari menu.
+    """
+    print("""
+    Terima kasih..
+    """)
     
 
 def menu():
@@ -240,53 +332,53 @@ def menu():
 7. Check Order
 8. Check Out Order
 ''')
-    print("Selamat Datang di Toko")
     menu = input("Choose a menu: ")
-    match menu:
-        case "1":
-            item_name = input("Enter the item name: ")
-            item_qty = float(input("Enter the item quantity: ") or 0)
-            item_price = float(input("Enter the item price: ") or 0)
 
-            add_item(item_name,item_qty,item_price)
+    if(menu=='1'):
+        item_name = input("Enter the item name: ")
+        item_qty = float(input("Enter the item quantity: ") or 0)
+        item_price = float(input("Enter the item price: ") or 0)
 
-        case "2":
-            item_name = input("Enter the item name: ")
-            new_item_name = input("Enter the new item name: ")
+        add_item(item_name,item_qty,item_price)
 
-            update_item_name(item_name, new_item_name)
+    elif(menu=='2'):
+        item_name = input("Enter the item name: ")
+        new_item_name = input("Enter the new item name: ")
 
-        case "3":
-            item_name = input("Enter the item name:")
-            new_item_qty = input("Enter the new item quantity: ")
+        update_item_name(item_name, new_item_name)
 
-            update_item_qty(item_name, new_item_qty)
+    elif(menu=='3'):
+        item_name = input("Enter the item name:")
+        new_item_qty = float(input("Enter the new item quantity: ") or 0)
 
-        case "4":
-            item_name = input("Enter the item name:")
-            new_item_price = input("Enter the new item price: ")
+        update_item_qty(item_name, new_item_qty)
 
+    elif(menu=='4'):
+        item_name = input("Enter the item name:")
+        new_item_price = float(input("Enter the new item price: ") or 0)
 
-            update_item_price(item_name, new_item_price)
+        update_item_price(item_name, new_item_price)
 
-        case "5":
-            item_name = input("Enter the item name you want to delete:")
+    elif(menu=='5'):
+        item_name = input("Enter the item name you want to delete:")
 
-            delete_item(item_name)
+        delete_item(item_name)
 
-        case "6":
-           reset_transaction()
+    elif(menu=='6'):
+        reset_transaction()
 
-        case "7":
-            check_order()
+    elif(menu=='7'):
+        check_order()
 
-        case "8":
-            check_out()
+    elif(menu=='8'):
+        check_out()
 
-        case _:
-            print("Invalid input.")
+    elif(menu=='9'):
+        leave()
+    # if user input other than available in menu
+    else:
+        menu()
     
-
 menu()
